dog-breed-classifier/pyproject.toml
[project]
name = "dog-breed-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]

[tool.ruff]
line-length = 95
fix = true
exclude = ["*.ipynb"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"
docstring-code-format = true
